refactor(street-check): split ui/tab.ts into format, dom and settings panel
tab.ts held 1063 lines mixing pure presentation helpers, DOM builders, the
rendering and diffing loop, viewport filtering, WME selection sync, and a
157-line settings panel. It also forced edit-panel.ts to import the whole tab
module to reach four small formatting functions.

Three modules extracted, no behaviour change:

- ui/format.ts: the nine already-exported, already-tested pure helpers
  (LEGEND_KEYS, STATE_KEYS, isDarkBackground, statusEmoji, formatNote,
  IssueGroup, groupIssues, geometryIntersectsBbox, bboxOfIssues). edit-panel.ts
  and the two test files now import from here instead of tab.ts.
- ui/dom.ts: el, toggleSwitch and buildSubsection, which touched no class state.
- ui/settings-panel.ts: buildSettings, now buildSettingsPanel(ctx) with an
  explicit five-field context. Only what genuinely needs TabUI internals is
  injected: rebuild (language change) and viewportOnlyToggle (the switch is
  duplicated in the master row and both instances must track each other).

wmeThemeIsDark stays in tab.ts: it walks the pane's ancestry and belongs to the
rendering side.

tab.ts: 1063 -> 716 lines.

The move was made with a script rather than retyped, and verified by normalising
the this/ctx renames and comparing the extracted body against the original: the
5045-character normalised forms are identical. This matters because no test
covers the settings panel, so the compiler and that comparison are the only
guards. Typecheck, 294 tests and the build are green.
