+++ src/street-name-checker/ui/tab.test.ts
@@ -1,7 +1,7 @@
 import type { LineString } from "geojson";
 import { describe, expect, it } from "vitest";
 import type { Issue, IssueStatus } from "../matching/evaluate";
-import { bboxOfIssues, formatNote, groupIssues, isDarkBackground } from "../ui/tab";
+import { bboxOfIssues, formatNote, groupIssues, isDarkBackground } from "../ui/format";
 import { setLocale } from "../i18n";
 
 const GEOMETRY: LineString = {
